Move AugeasConfigurator to letsencrypt_apache.
https://github.com/letsencrypt/lets-encrypt-preview/pull/531#issuecomment-114285541
