--- docs/api/augeas_configurator.rst
@@ -1,5 +1,0 @@
-:mod:`letsencrypt.augeas_configurator`
---------------------------------------
-
-.. automodule:: letsencrypt.augeas_configurator
-   :members:
