+++ docs/pkgs/letsencrypt_apache.rst
@@ -27,3 +27,10 @@
 
 .. automodule:: letsencrypt_apache.parser
    :members:
+
+
+:mod:`letsencrypt_apache.augeas_configurator`
+=============================================
+
+.. automodule:: letsencrypt_apache.augeas_configurator
+   :members:
